+++ Sources/SwiftUIKit/Pickers/SystemFontListPicker.swift
@@ -24,26 +24,22 @@ public struct SystemFontListPicker: View {
      Create a font list picker.
      
      - Parameters:
-       - title: The list's navigation title.
        - selectedFontName: The selected font name.
        - fonts: The fonts to display in the list, by default `all` with the selected font topmost.
        - itemFontSize: The font size to use in the list items.
        - dismissAfterPick: Whether or not to dismiss the picker after a font has been selected, by default `true`.
      */
     public init(
-        title: String,
         selectedFontName: Binding<String>,
         fonts: [SystemFontPickerFont]? = nil,
         itemFontSize: CGFloat = 20,
         dismissAfterPick: Bool = true) {
-        self.title = title
         self._selectedFontName = selectedFontName
         self.fonts = fonts ?? .all.moveTopmost(selectedFontName.wrappedValue)
         self.itemFontSize = itemFontSize
         self.dismissAfterPick = dismissAfterPick
     }
     
-    private let title: String
     private let itemFontSize: CGFloat
     private let dismissAfterPick: Bool
     
@@ -55,7 +51,6 @@ public struct SystemFontListPicker: View {
      */
     private let fonts: [SystemFontPickerFont]
     
-    
     public var body: some View {
         let font = Binding(
             get: { SystemFontPickerFont(fontName: selectedFontName) },
@@ -63,7 +58,6 @@ public struct SystemFontListPicker: View {
         )
         
         ListPicker(
-            title: title,
             items: fonts,
             selection: font,
             dismissAfterPick: dismissAfterPick) { font, isSelected in
@@ -75,6 +69,17 @@ public struct SystemFontListPicker: View {
     }
 }
 
+private extension View {
+    
+    @ViewBuilder
+    func withTitle(_ title: String) -> some View {
+        #if os(iOS) || os(tvOS) || os(watchOS)
+        self.navigationBarTitle(title)
+        #else
+        self
+        #endif
+    }
+}
 
 struct SystemFontListPicker_Previews: PreviewProvider {
     
@@ -85,8 +90,8 @@ struct SystemFontListPicker_Previews: PreviewProvider {
         var body: some View {
             NavigationView {
                 SystemFontListPicker(
-                    title: "Pick a font",
                     selectedFontName: $font)
+                .withTitle("Pick a font")
             }
         }
     }